+++ playwright.config.ts
@@ -15,28 +15,34 @@ export default defineConfig({
 	use: {
 		baseURL: process.env.CI ? 'http://localhost:4173' : 'http://localhost:5173',
 		trace: 'on-first-retry',
-		screenshot: 'only-on-failure'
+		screenshot: 'only-on-failure',
+		launchOptions: {
+			args: [
+				'--use-gl=angle',
+				'--disable-gpu-sandbox',
+				...(process.env.CI ? ['--disable-gpu', '--no-sandbox', '--disable-setuid-sandbox'] : [])
+			]
+		}
 	},
 
-	projects: process.env.CI
-		? [
-				{
-					name: 'chromium',
-					use: {
-						...devices['Desktop Chrome'],
-						launchOptions: {
-							args: ['--use-gl=swiftshader', '--no-sandbox']
-						}
-					}
-				},
-				{ name: 'webkit', use: devices['Desktop Safari'] }
-			]
-		: [{ name: 'chromium', use: devices['Desktop Chrome'] }],
+	projects: [
+		{
+			name: 'chromium',
+			use: devices['Desktop Chrome']
+		},
+		{
+			name: 'firefox',
+			use: devices['Desktop Firefox']
+		},
+		{
+			name: 'webkit',
+			use: devices['Desktop Safari']
+		}
+	],
 
 	webServer: {
 		command: process.env.CI ? 'bun run preview' : 'bun run dev',
 		url: process.env.CI ? 'http://localhost:4173' : 'http://localhost:5173',
-		reuseExistingServer: !process.env.CI,
-		timeout: 2 * 60 * 1000
+		reuseExistingServer: !process.env.CI
 	}
 })